+++ src/constants/colors/numerical-color-sun.ts
@@ -1,0 +1,1 @@
+export const NUMERICAL_COLOR_SUN = 16376750;