+++ src/features/discord/guilds/services/discord-guild-sonia.service.ts
@@ -2,6 +2,7 @@ import { Guild, GuildChannel } from "discord.js";
 import _ from "lodash";
 import { filter, take } from "rxjs/operators";
 import { AbstractService } from "../../../../classes/services/abstract.service";
+import { ONE_EMITTER } from "../../../../constants/one-emitter";
 import { ServiceNameEnum } from "../../../../enums/service-name.enum";
 import { wrapInQuotes } from "../../../../functions/formatters/wrap-in-quotes";
 import { ChalkService } from "../../../logger/services/chalk/chalk.service";
@@ -141,7 +142,7 @@ export class DiscordGuildSoniaService extends AbstractService {
         filter((isReady: Readonly<boolean>): boolean =>
           _.isEqual(isReady, true)
         ),
-        take(1)
+        take(ONE_EMITTER)
       )
       .subscribe({
         next: (): void => {