+++ src/features/logger/constants/levels/logger-debug-level.ts
@@ -1,0 +1,1 @@
+export const LOGGER_DEBUG_LEVEL = 4;